Extract Router to new file
